drm/mediatek: dsi: Do not restore mode on sucessful host transfer
The DSI controller on MT6735 needs some time between successive command-video-command
mode switches, and therefore currently fails to switch back to command mode after
a host transfer followed by a video mode switch. These successive mode switches mainly
happen in mtk_dsi_host_transfer, where the mode is restored after performing a transfer,
resulting in mode switches on each transfer in a sequence (panel initialization, for
example), with the time between a switch back to video mode and a switch to command mode
for a new transfer being separated by just a few microseconds. To avoid this, leave the
DSI controller in command mode after a successful host transfer since it is very likely
that there will be more transfers upcoming after one. A call to mtk_dsi_output_enable
sets the appropriate mode for panel operation, so restoring the mode on host transfers
is not strictly necessary.

Mode switches should not be necessary for host transfers since the DSI controller
supports sending commands while in video mode, but that is not implemented at the
moment, so this is the only way to roll for now.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>